writing: relocate detection engine out of hooks/
Move tropes, scan, and wordlists into a detection/ module so hooks/,
skills/scan, skills/rewrite, and skills/analyze become peer consumers
instead of reaching through ../../../hooks/. Split markdown.ts into
detection/paths.ts (file classification) and hooks/io.ts (hook I/O).

Add regexCatalog() to the engine so the structural audit consumes a
stable interface rather than filtering and globalizing the raw PATTERNS
array. structural.ts loses ~30 lines of engine-internal knowledge.
