+++ plugins/writing/detection/paths.ts
@@ -1,10 +1,4 @@
 import { resolve } from "node:path";
-import type { SyncHookJSONOutput } from "@anthropic-ai/claude-agent-sdk";
-
-export type { SyncHookJSONOutput };
-
-export type WriteInput = { file_path: string; content: string };
-export type EditInput = { file_path: string; new_string: string };
 
 export function getExtension(filePath: string): string {
   const parts = filePath.split(".");
@@ -30,23 +24,4 @@ const PROSE_EXTENSIONS = new Set(["md", "markdown", "txt", "mdx", "rst", "adoc"]
 
 export function isProseFile(filePath: string): boolean {
   return PROSE_EXTENSIONS.has(getExtension(filePath));
-}
-
-export function formatDecision(decision: "deny" | "ask", reason: string): SyncHookJSONOutput {
-  return {
-    hookSpecificOutput: {
-      hookEventName: "PreToolUse",
-      permissionDecision: decision,
-      permissionDecisionReason: reason,
-    },
-  };
-}
-
-export function formatContext(context: string): SyncHookJSONOutput {
-  return {
-    hookSpecificOutput: {
-      hookEventName: "PreToolUse",
-      additionalContext: context,
-    },
-  };
 }
