+++ plugins/writing/detection/scan.test.ts
