+++ plugins/writing/hooks/io.ts
@@ -1,0 +1,25 @@
+import type { SyncHookJSONOutput } from "@anthropic-ai/claude-agent-sdk";
+
+export type { SyncHookJSONOutput };
+
+export type WriteInput = { file_path: string; content: string };
+export type EditInput = { file_path: string; new_string: string };
+
+export function formatDecision(decision: "deny" | "ask", reason: string): SyncHookJSONOutput {
+  return {
+    hookSpecificOutput: {
+      hookEventName: "PreToolUse",
+      permissionDecision: decision,
+      permissionDecisionReason: reason,
+    },
+  };
+}
+
+export function formatContext(context: string): SyncHookJSONOutput {
+  return {
+    hookSpecificOutput: {
+      hookEventName: "PreToolUse",
+      additionalContext: context,
+    },
+  };
+}
